[agent-base] Allow for never relying on stack trace (#170)
During an HTTP request, if `opts.protocol` is set, then use that to
determine the `secureEndpoint` value. Coupling that with an explicitly
set `agent.protocol` in the instance (essentially "locking" it for use
with either `http` or `https` module), then the agent will _never_ rely
on the stack trace. By default it will still rely on the stack trace to
allow for agent instances to be used with either `http` or `https`
modules.

Also moved the internal props to an object using a symbol, instead of
relying on underscore props.

Fixes #158.